Fix for Hua's review
