+++ device-telemetry/Services/Models/StatusResultServiceModel.cs
@@ -11,13 +11,7 @@ namespace Microsoft.Azure.IoTSolutions.DeviceTelemetry.Services.Models
 
         [JsonProperty(PropertyName = "Message")]
         public string Message { get; set; }
-
-        public StatusResultServiceModel(StatusResultServiceModel result)
-        {
-            IsHealthy = result.IsHealthy;
-            Message = result.Message;
-        }
-
+        
         [JsonConstructor]
         public StatusResultServiceModel(bool isHealthy, string message)
         {
